New Version
This is the newest working version of the game. In this step we have added the background as a space, a player as a red car with gun on it, some flying enemies(aliens) that spawn every constant time. The music plays as you start the game. Score and lifes are also added in this version.